+++ backend/apps/notifications/apps.py
@@ -4,3 +4,6 @@ from django.apps import AppConfig
 class NotificationsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'apps.notifications'
+
+    def ready(self):
+        import apps.notifications.signals  # noqa
